added wasm support
